+++ .github/workflows/flutter.yml
@@ -1,28 +1,28 @@
-name: flutter
-
-on:
-  push:
-    branches: [ master ]
-  pull_request:
-    branches: [ master ]
-
-jobs: 
-  build:
-    name: build on ${{ matrix.os }}
-    runs-on: ${{ matrix.os }}
-    strategy:
-      matrix:
-        os: [ubuntu-latest, windows-latest, macos-latest]
-    steps:
-      - uses: actions/checkout@v1
-      - uses: actions/setup-java@v1
-        with:
-          java-version: '12.x'
-      - uses: subosito/flutter-action@v1
-        with:
-          channel: 'stable' 
-      - run: flutter pub get
-#      - run: flutter test
-      - run: flutter build apk
-      - run: flutter build web 
-      - run: flutter build ios
+#name: flutter
+#
+#on:
+#  push:
+#    branches: [ master ]
+#  pull_request:
+#    branches: [ master ]
+#
+#jobs:
+#  build:
+#    name: build on ${{ matrix.os }}
+#    runs-on: ${{ matrix.os }}
+#    strategy:
+#      matrix:
+#        os: [ubuntu-latest, windows-latest, macos-latest]
+#    steps:
+#      - uses: actions/checkout@v1
+#      - uses: actions/setup-java@v1
+#        with:
+#          java-version: '12.x'
+#      - uses: subosito/flutter-action@v1
+#        with:
+#          channel: 'stable'
+#      - run: flutter pub get
+##      - run: flutter test
+#      - run: flutter build apk
+#      - run: flutter build web
+#      - run: flutter build ios
